allow force-auth/passive per request

+++ pac4j-saml/src/main/java/org/pac4j/saml/profile/api/SAML2ResponseValidator.java
@@ -20,5 +20,5 @@ public interface SAML2ResponseValidator {
      */
     Credentials validate(SAML2MessageContext context);
 
-    void setAcceptedSkew(int acceptedSkew);
+    void setAcceptedSkew(long acceptedSkew);
 }
